인게임 내 건설 UI 기능 연결 완료
- 인게임에서 상점 구매 -> 빈 땅 클릭 -> 작물 건설 기능 추가 완료
- 낚시 미니게임의 오류 수정(노트 이탈 문제 추가 해결, 버튼이 계속 눌리는 문제 해결, 남은 시간의 소수점 이상의 수만 출력하도록 변경)

+++ Assets/MiniGames/FishingGame/Script/Manager/TimeManager.cs
@@ -17,21 +17,24 @@ public class TimeManager : MonoBehaviour
     {
         _gameOver.SetActive(false);
     }
-    void Update () {
+    void Update()
+    {
 
-if(done>0F){
+        if (done > 0F)
+        {
 
-done-=Time.deltaTime; //시간 -1
+            done -= Time.deltaTime; //시간 -1
 
-_time.text = ""+done;
+            _time.text = "" + (int)done;
 
-}
+        }
 
-else{
+        else
+        {
 
-_gameOver.SetActive(true);
+            _gameOver.SetActive(true);
 
-}
+        }
